src/servers/hallucinationGrader/requirements.txt
fastapi
uvicorn
pydantic
langchain-core
python-dotenv
langchain-ollama
langgraph
ipython
langchain